Removed BettingSys, Started Crazy8 Implementation
