+++ .kiro/specs/crazy-8-game/requirements.md
@@ -1,0 +1,132 @@
+# Requirements Document
+
+## Introduction
+
+This document outlines the requirements for implementing a Crazy 8 card game as an addition to the existing Blackjack multiplayer platform. Crazy 8 is a classic card game that uses a standard 52-card deck where players try to be the first to play all their cards by matching suit or rank, with 8s serving as wild cards.
+
+## Requirements
+
+### Requirement 1: Game Setup and Initialization
+
+**User Story:** As a player, I want to start a new Crazy 8 game so that I can play with friends online.
+
+#### Acceptance Criteria
+
+1. WHEN a player selects Crazy 8 from the game menu THEN the system SHALL display game mode options (2-6 players)
+2. WHEN a player creates a Crazy 8 room THEN the system SHALL generate a unique room code
+3. WHEN players join a Crazy 8 room THEN the system SHALL display waiting lobby with player list
+4. WHEN the room creator starts the game THEN the system SHALL deal 7 cards to each player
+5. WHEN cards are dealt THEN the system SHALL place one card face-up as the starting discard pile
+6. IF the starting card is an 8 THEN the system SHALL reshuffle and draw a new starting card
+
+### Requirement 2: Basic Gameplay Mechanics
+
+**User Story:** As a player, I want to play cards according to Crazy 8 rules so that I can participate in the game.
+
+#### Acceptance Criteria
+
+1. WHEN it's a player's turn THEN the system SHALL highlight playable cards in their hand
+2. WHEN a player plays a valid card THEN the system SHALL move the card to the discard pile
+3. WHEN a player plays a card matching suit OR rank THEN the system SHALL accept the play
+4. WHEN a player plays an 8 THEN the system SHALL prompt them to choose a new suit
+5. WHEN a player cannot play any card THEN the system SHALL require them to draw from the deck
+6. WHEN a player draws a playable card THEN the system SHALL allow them to play it immediately OR pass
+7. WHEN the draw pile is empty THEN the system SHALL reshuffle the discard pile (except top card)
+
+### Requirement 3: Turn Management
+
+**User Story:** As a player, I want turns to progress in order so that everyone gets a fair chance to play.
+
+#### Acceptance Criteria
+
+1. WHEN a game starts THEN the system SHALL randomly select the first player
+2. WHEN a player completes their turn THEN the system SHALL advance to the next player clockwise
+3. WHEN a player takes too long (60 seconds) THEN the system SHALL automatically make them draw a card
+4. WHEN it's a player's turn THEN the system SHALL display a clear turn indicator
+5. WHEN it's not a player's turn THEN the system SHALL disable their card interactions
+
+### Requirement 4: Wild Card (8s) Functionality
+
+**User Story:** As a player, I want to use 8s as wild cards so that I can change the active suit strategically.
+
+#### Acceptance Criteria
+
+1. WHEN a player plays an 8 THEN the system SHALL display a suit selection interface
+2. WHEN a player selects a new suit THEN the system SHALL update the active suit for the next player
+3. WHEN an 8 is played THEN the system SHALL visually indicate the chosen suit on the discard pile
+4. WHEN the next player's turn begins THEN they SHALL be able to play cards of the chosen suit OR any rank 8
+
+### Requirement 5: Win Condition and Scoring
+
+**User Story:** As a player, I want to win by playing all my cards first so that I can be declared the winner.
+
+#### Acceptance Criteria
+
+1. WHEN a player plays their last card THEN the system SHALL declare them the winner
+2. WHEN a player has one card remaining THEN the system SHALL display a visual indicator
+3. WHEN the game ends THEN the system SHALL calculate points for remaining cards in other players' hands
+4. WHEN points are calculated THEN face cards SHALL count as 10, 8s as 50, and number cards as face value
+5. WHEN the round ends THEN the system SHALL display final scores and offer to play again
+
+### Requirement 6: User Interface and Experience
+
+**User Story:** As a player, I want an intuitive interface so that I can easily understand the game state and make moves.
+
+#### Acceptance Criteria
+
+1. WHEN viewing the game THEN the system SHALL display my hand horizontally at the bottom
+2. WHEN viewing the game THEN the system SHALL show the current discard pile card prominently
+3. WHEN viewing the game THEN the system SHALL display other players' card counts around the table
+4. WHEN it's my turn THEN the system SHALL highlight playable cards with a glow effect
+5. WHEN hovering over a card THEN the system SHALL show a preview of the card enlarged
+6. WHEN the active suit changes THEN the system SHALL display the current suit clearly
+7. WHEN cards are played THEN the system SHALL show smooth animations
+
+### Requirement 7: Multiplayer Integration
+
+**User Story:** As a player, I want to play with multiple friends online so that we can enjoy the game together.
+
+#### Acceptance Criteria
+
+1. WHEN creating a game THEN the system SHALL support 2-6 players
+2. WHEN players join/leave THEN the system SHALL update all clients in real-time
+3. WHEN a player disconnects THEN the system SHALL pause the game for 60 seconds
+4. WHEN a disconnected player returns THEN the system SHALL restore their game state
+5. WHEN a player permanently leaves THEN the system SHALL remove them and continue with remaining players
+6. WHEN only one player remains THEN the system SHALL end the game
+
+### Requirement 8: Game State Persistence
+
+**User Story:** As a player, I want the game to handle interruptions gracefully so that we can resume play.
+
+#### Acceptance Criteria
+
+1. WHEN a game is in progress THEN the system SHALL maintain complete game state on the server
+2. WHEN a player refreshes their browser THEN the system SHALL restore their view of the game
+3. WHEN network issues occur THEN the system SHALL attempt to reconnect automatically
+4. WHEN reconnection succeeds THEN the system SHALL sync the current game state
+5. WHEN the server restarts THEN active games SHALL be preserved (if possible)
+
+### Requirement 9: Integration with Existing Platform
+
+**User Story:** As a user, I want Crazy 8 to integrate seamlessly with the existing game platform so that I have a consistent experience.
+
+#### Acceptance Criteria
+
+1. WHEN accessing the main menu THEN the system SHALL display both Blackjack and Crazy 8 options
+2. WHEN switching between games THEN the system SHALL maintain the same UI/UX patterns
+3. WHEN using the room system THEN Crazy 8 SHALL use the same room codes and joining mechanism
+4. WHEN viewing player information THEN the system SHALL use consistent styling and layout
+5. WHEN handling errors THEN the system SHALL use the same error handling and display patterns
+
+### Requirement 10: Performance and Scalability
+
+**User Story:** As a player, I want the game to run smoothly so that I can enjoy uninterrupted gameplay.
+
+#### Acceptance Criteria
+
+1. WHEN playing with 6 players THEN the system SHALL maintain responsive performance (<100ms response time)
+2. WHEN multiple games are running THEN the system SHALL handle concurrent rooms efficiently
+3. WHEN cards are animated THEN the system SHALL maintain 60fps performance
+4. WHEN the game state updates THEN the system SHALL only send necessary data to minimize bandwidth
+5. WHEN memory usage grows THEN the system SHALL clean up completed games automatically
