+++ .kiro/specs/crazy-8-game/tasks.md
@@ -1,0 +1,231 @@
+# Implementation Plan
+
+## Phase 1: Foundation and Core Setup (3-4 days)
+
+- [x] 1. Create game mode selection system
+
+
+
+  - Create GameModeSelector component in frontend
+  - Add routing logic to switch between Blackjack and Crazy 8
+  - Update main App component to handle multiple game modes
+  - _Requirements: 9.1, 9.3_
+
+- [x] 2. Set up Crazy 8 game types and interfaces
+
+
+
+
+
+  - Define Crazy8Player, Crazy8GameState, and related TypeScript interfaces
+  - Create Crazy8Room type extending existing Room interface
+  - Add game type discrimination to room management
+  - _Requirements: 7.1, 8.1_
+
+- [ ] 3. Create basic Crazy8GameService structure
+  - Implement Crazy8GameService class with core method signatures
+  - Set up game state initialization and management
+  - Create basic room creation for Crazy 8 games
+  - _Requirements: 1.2, 1.3_
+
+- [ ] 4. Implement card dealing and initial game setup
+  - Create dealInitialHand method (7 cards per player)
+  - Implement starting discard pile setup
+  - Add validation to prevent 8s as starting card
+  - _Requirements: 1.4, 1.5, 1.6_
+
+## Phase 2: Core Game Logic (4-5 days)
+
+- [ ] 5. Implement basic card validation system
+  - Create isValidPlay function for suit/rank matching
+  - Implement getPlayableCards function
+  - Add server-side move validation
+  - _Requirements: 2.2, 2.3_
+
+- [ ] 6. Create turn management system
+  - Implement turn advancement logic
+  - Add current player tracking and validation
+  - Create turn timeout mechanism (60 seconds)
+  - _Requirements: 3.1, 3.2, 3.3_
+
+- [ ] 7. Implement draw card functionality
+  - Create drawCard method in game service
+  - Handle empty deck reshuffling logic
+  - Add option to play drawn card immediately
+  - _Requirements: 2.5, 2.6, 2.7_
+
+- [ ] 8. Add basic win condition detection
+  - Implement checkWinCondition method
+  - Create game ending logic when player empties hand
+  - Add winner announcement system
+  - _Requirements: 5.1, 5.2_
+
+## Phase 3: Wild Card (8s) Implementation (2-3 days)
+
+- [ ] 9. Create suit selection system
+  - Implement SuitSelector component
+  - Add suit selection modal/overlay
+  - Create chooseSuit method in game service
+  - _Requirements: 4.1, 4.2_
+
+- [ ] 10. Implement 8s wild card logic
+  - Add special handling for 8s in card validation
+  - Implement active suit tracking and updates
+  - Create visual indicators for chosen suit
+  - _Requirements: 4.3, 4.4_
+
+- [ ] 11. Handle 8s in game flow
+  - Integrate suit selection into turn flow
+  - Add validation for 8s being playable anytime
+  - Update game state broadcasting for suit changes
+  - _Requirements: 2.4, 4.1_
+
+## Phase 4: User Interface Development (5-6 days)
+
+- [ ] 12. Create Crazy8Table component
+  - Build main game table layout
+  - Position players around the table
+  - Add center area with deck and discard pile
+  - _Requirements: 6.1, 6.2, 6.3_
+
+- [ ] 13. Implement PlayerHand component
+  - Create horizontal hand display at bottom
+  - Add card interaction and selection
+  - Implement playable card highlighting
+  - _Requirements: 6.1, 6.4_
+
+- [ ] 14. Create DiscardPile and DeckPile components
+  - Display current top card prominently
+  - Show active suit indicator
+  - Add draw pile with card count
+  - _Requirements: 6.2, 6.6_
+
+- [ ] 15. Add OtherPlayerView components
+  - Display other players' card counts
+  - Show current turn indicators
+  - Add player status and positioning
+  - _Requirements: 6.3, 3.4_
+
+- [ ] 16. Implement card animations and effects
+  - Add smooth card play animations
+  - Create card hover and selection effects
+  - Implement turn transition animations
+  - _Requirements: 6.5, 6.7_
+
+## Phase 5: Scoring and Game Completion (2-3 days)
+
+- [ ] 17. Implement scoring system
+  - Create calculateCardPoints function
+  - Implement end-game scoring calculation
+  - Add score display and tracking
+  - _Requirements: 5.3, 5.4_
+
+- [ ] 18. Add game completion flow
+  - Create end game screen with scores
+  - Implement play again functionality
+  - Add winner celebration effects
+  - _Requirements: 5.5_
+
+- [ ] 19. Create game statistics and history
+  - Track wins/losses per player
+  - Add round history display
+  - Implement score accumulation across rounds
+  - _Requirements: 5.4, 5.5_
+
+## Phase 6: Multiplayer Integration and Polish (3-4 days)
+
+- [ ] 20. Enhance multiplayer support
+  - Support 2-6 players in rooms
+  - Add dynamic player positioning
+  - Implement player join/leave during lobby
+  - _Requirements: 7.1, 7.2_
+
+- [ ] 21. Implement disconnection handling
+  - Add player disconnect detection
+  - Create reconnection logic with state restoration
+  - Handle permanent player removal
+  - _Requirements: 7.3, 7.4, 7.5_
+
+- [ ] 22. Add game state persistence
+  - Implement browser refresh recovery
+  - Add automatic reconnection on network issues
+  - Create game state synchronization
+  - _Requirements: 8.1, 8.2, 8.3, 8.4_
+
+- [ ] 23. Optimize performance and networking
+  - Minimize game state broadcast payload
+  - Implement efficient card rendering
+  - Add memory cleanup for finished games
+  - _Requirements: 10.1, 10.2, 10.4, 10.5_
+
+## Phase 7: Testing and Quality Assurance (2-3 days)
+
+- [ ] 24. Create comprehensive unit tests
+  - Test card validation logic
+  - Test turn management system
+  - Test scoring calculations
+  - Test win condition detection
+
+- [ ] 25. Implement integration tests
+  - Test complete game flow
+  - Test multiplayer synchronization
+  - Test error handling and recovery
+  - Test performance under load
+
+- [ ] 26. Add end-to-end testing
+  - Test full game sessions with multiple players
+  - Test disconnection and reconnection scenarios
+  - Test edge cases and error conditions
+  - Validate UI responsiveness and animations
+
+## Phase 8: Final Integration and Deployment (1-2 days)
+
+- [ ] 27. Integrate with existing platform
+  - Ensure consistent UI/UX with Blackjack
+  - Test room system compatibility
+  - Validate error handling consistency
+  - _Requirements: 9.2, 9.4, 9.5_
+
+- [ ] 28. Performance optimization and cleanup
+  - Profile and optimize critical paths
+  - Clean up unused code and dependencies
+  - Optimize bundle size and loading times
+  - _Requirements: 10.3_
+
+- [ ] 29. Final testing and bug fixes
+  - Conduct user acceptance testing
+  - Fix any remaining bugs or issues
+  - Validate all requirements are met
+  - Prepare for production deployment
+
+## Estimated Timeline: 3-4 weeks total
+
+### Week 1: Foundation + Core Logic (Phases 1-2)
+- Game mode selection
+- Basic game service
+- Card validation and turns
+
+### Week 2: Wild Cards + UI (Phases 3-4)
+- 8s implementation
+- Main game interface
+- Card interactions
+
+### Week 3: Scoring + Multiplayer (Phases 5-6)
+- Game completion
+- Advanced multiplayer features
+- Performance optimization
+
+### Week 4: Testing + Integration (Phases 7-8)
+- Comprehensive testing
+- Final integration
+- Production readiness
+
+## Success Criteria
+
+- [ ] All requirements from requirements.md are implemented and tested
+- [ ] Game supports 2-6 players smoothly
+- [ ] UI is intuitive and responsive
+- [ ] No major bugs or performance issues
+- [ ] Seamless integration with existing Blackjack platform
+- [ ] Complete game sessions work end-to-end
+- [ ] Proper error handling and recovery mechanisms
