+++ backend/src/services/multiplayer/multiplayerGameService.ts
@@ -2,7 +2,8 @@ import { Socket, Server } from 'socket.io';
 import { Card } from '../../types/gameTypes';
 import { DealingService } from '../game/dealingService';
 import { GameLogic } from '../game/gameLogic';
-import { validateBet } from '../../utils/validationUtils';
+import { calculateHand } from '../../utils/cardUtils';
+
 
 /**
  * Service for managing multiplayer game state
@@ -123,7 +124,7 @@ export class MultiplayerGameService {
    * Calculates the dealer's hand total
    */
   private calculateDealerTotal(hand: Card[]): number {
-    return GameLogic.calculateHand(hand).total;
+    return calculateHand(hand).total;
   }
   
   /**
@@ -147,16 +148,9 @@ export class MultiplayerGameService {
         result = GameLogic.determineWinner(player.hand, dealerHand);
       }
       
-      const payout = GameLogic.calculatePayout(player.bet, result, player.isBlackjack);
-      
       gameState.results[player.id] = {
-        result,
-        payout,
-        finalBalance: player.balance + payout
+        result
       };
-      
-      // Update player balance
-      player.balance += payout;
     });
   }
