+++ backend/src/types/index.ts
@@ -1,0 +1,3 @@
+// Game types exports
+export * from './gameTypes';
+export * from './crazy8Types';
