+++ backend/src/utils/index.ts
@@ -2,3 +2,4 @@
 export * from './cardUtils';
 export * from './validationUtils';
 export * from './roomUtils';
+export * from './typeGuards';
