+++ frontend/src/components/index.js
@@ -1,3 +1,5 @@
 export { default as CasinoTable } from './CasinoTable';
 export { default as PlayerPosition } from './PlayerPosition';
 export { default as BettingPanel } from './BettingPanel';
+export { default as GameModeSelector } from './GameModeSelector';
+export { default as BlackjackGame } from './BlackjackGame';
